+++ CHANGES.md
@@ -3,6 +3,7 @@
 ### Last Version
 
   * Store reflection objects separate from it's definition.
+  * [BC BREAK] Change `registerFactory` arguments order.
 
 ### 0.1.0 (2014-10-03)